requirements/commont.txt
django==5.1.1
django-environ==0.11.2
psycopg2==2.9.9
Pillow==10.4.0
djangorestframework==3.15.2

django-extensions==3.2.3

faker==30.0.0
pytest==8.3.3
pytest-django==4.9.0
pytest-testdox==3.1.0
factory-boy==3.3.1
